+++ packages/auth/src/resolvers/evmVerifyChallenge.ts
@@ -1,5 +1,5 @@
 import { maybe, toCamelCase } from '@moralisweb3/core';
-import { EvmAddress, EvmChain } from '@moralisweb3/evm-utils';
+import { EvmAddress, EvmChain } from '@moralisweb3/common-evm-utils';
 import { createEndpoint, createEndpointFactory } from '@moralisweb3/api-utils';
 import { operations } from '../generated/types';