+++ packages/solUtils/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @moralisweb3/sol-utils
+# @moralisweb3/common-sol-utils
 
 ## 2.6.6